requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
oauthlib==3.1.0
PySocks==1.7.1
python-decouple==3.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
urllib3==1.26.2
wincertstore==0.2
